integrators/other/ftp-sftp/v1.0.0/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
aiokafka==0.12.0
aiosqlite==0.20.0
prometheus-client==0.21.1
pyyaml==6.0.2
aioftp==0.22.3
asyncssh==2.18.0
